requirements.txt
Requests
Pillow
RPi.GPIO
spidev